+++ tests/datatypes/tests/dataset/test_revise_strategy.py
@@ -1,42 +1,127 @@
+from unittest.mock import patch
+
 from django.test import override_settings
 
+from datagrowth.datatypes.datasets.constants import GrowthState, GrowthStrategy
+from datagrowth.datatypes.documents.tasks.document import dispatch_data_storage_tasks
+
 from datatypes.tests.dataset import base as test_cases
-from datatypes.models import Dataset, DatasetVersion
+from datatypes.models import Dataset, Document
 from resources.models import EntityListResource
 
 
-class TestInitialDatasetReviseStrategy(test_cases.BaseDatasetTestCase):
+DISPATCH_DATA_STORAGE_TARGET = "datagrowth.datatypes.documents.tasks.document.dispatch_data_storage_tasks"
+
+
+class TestInitialDatasetReviseStrategy(test_cases.InitialDatasetGrowthTestCase):
 
     dataset_model = Dataset
     signature = "setting1=const&test"
     entity_type = "paper"
 
-    def test_growth_success(self):
+
+class TestContinuationDatasetReviseStrategy(test_cases.BaseDatasetGrowthTestCase):
+
+    dataset_model = Dataset
+    signature = "setting1=const&test"
+    entity_type = "paper"
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        cls.create_historic_dataset_version()
+
+    @patch(DISPATCH_DATA_STORAGE_TARGET, side_effect=dispatch_data_storage_tasks)
+    @patch.object(EntityListResource, "PARAMETERS", {"size": 3, "page_size": 10})
+    def test_growth_success(self, dispatch_mock):
+        # Prepare test data
+        disappearing_title = "This title should disappear. It is overwritten by new entities"
+        remaining_title = "This title should remain. It is copied from old entities"
+        self.prepare_documents(properties={
+            "2": {
+                "title": disappearing_title
+            },
+            "4": {
+                "title": remaining_title
+            }
+        })
+        # Execute growth
         self.dataset.grow(self.entity_type, asynchronous=False)
-        self.assert_initial_grow_success(dataset_versions=1, collections=1, documents=20)
+        # Standard assertions
+        self.assert_grow_success(dataset_versions=2, collections=2, documents=40)
         self.assert_dataset_output(self.dataset, dataset_versions=1, collections=1, documents=20)
         self.assert_dataset_finish()
+        self.assertEqual(dispatch_mock.call_count, 0, "Did not expect Document tasks to get dispatched")
+        # Check if expected updates have taken place
+        self.assertEqual(
+            Document.objects.filter(properties__title=disappearing_title).count(), 1,
+            "Expected disappearing title to be preserved in old data, but overwritten in new data."
+        )
+        self.assertEqual(
+            Document.objects.filter(properties__title=remaining_title).count(), 2,
+            "Expected remaining title to be present in old and new data."
+        )
 
-    def test_growth_limit(self):
+    @patch(DISPATCH_DATA_STORAGE_TARGET, side_effect=dispatch_data_storage_tasks)
+    def test_growth_limit(self, dispatch_mock):
+        # Prepare test data
+        disappearing_title = "This title should disappear. It is overwritten by new entities"
+        remaining_title = "This title should remain. It is copied from old entities"
+        self.prepare_documents(properties={
+            "2": {
+                "title": disappearing_title
+            },
+            "5": {
+                "title": remaining_title
+            }
+        })
         # This test sets the limit to 3.
         # However the batch_size is 5 and therefor we expect to grow 5 Documents.
         self.dataset.grow(self.entity_type, asynchronous=False, limit=3)
-        self.assert_initial_grow_success(dataset_versions=1, collections=1, documents=5)
-        self.assert_dataset_output(self.dataset, dataset_versions=1, collections=1, documents=5)
+        # Standard assertions
+        self.assert_grow_success(dataset_versions=2, collections=2, documents=40)
+        self.assert_dataset_output(self.dataset, dataset_versions=1, collections=1, documents=20)
         self.assert_dataset_finish()
+        self.assertEqual(dispatch_mock.call_count, 0, "Did not expect Document tasks to get dispatched")
+        # Check if expected updates have taken place
+        self.assertEqual(
+            Document.objects.filter(properties__title=disappearing_title).count(), 1,
+            "Expected disappearing title to be preserved in old data, but overwritten in new data."
+        )
+        self.assertEqual(
+            Document.objects.filter(properties__title=remaining_title).count(), 2,
+            "Expected remaining title to be present in old and new data."
+        )
 
-    def test_seeding_error(self):
+    @patch(DISPATCH_DATA_STORAGE_TARGET, side_effect=dispatch_data_storage_tasks)
+    def test_seeding_error(self, dispatch_mock):
+        # Prepare test data
+        remaining_title = "This title should remain. It is copied from old entities and is never overwritten."
+        self.prepare_documents(properties={
+            "2": {
+                "title": remaining_title
+            }
+        })
+        # Execute growth
         self.dataset.grow("does_not_exist", asynchronous=False)
-        self.assert_initial_grow_success(dataset_versions=1, collections=1, documents=0)
-        self.assert_dataset_output(self.dataset, dataset_versions=1, collections=1, documents=0)
+        # Standard assertions
+        self.assert_grow_success(dataset_versions=2, collections=2, documents=40)
+        self.assert_dataset_output(self.dataset, dataset_versions=1, collections=1, documents=20)
+        self.assert_dataset_finish()
+        self.assertEqual(dispatch_mock.call_count, 0, "Did not expect Document tasks to get dispatched")
+        # Check if expected updates have taken place
+        self.assertEqual(
+            Document.objects.filter(properties__title=remaining_title).count(), 2,
+            "Expected remaining title to be present in old and new data."
+        )
         entity_list_resource = EntityListResource.objects.last()
-        dataset_version = DatasetVersion.objects.first()
+        dataset_version = self.dataset.get_current_dataset_version()
         self.assertEqual(dataset_version.errors, {
-            "tasks":  {
+            "tasks": {
                 "check_doi": {
                     "fail": 0,
                     "skipped": 0,
-                    "success": 0
+                    "success": 20
                 }
             },
             "seeding": {
@@ -48,14 +133,46 @@ class TestInitialDatasetReviseStrategy(test_cases.BaseDatasetTestCase):
                 }
             }
         })
-        self.assert_dataset_finish()
 
-    def test_task_error(self):
+    @patch(DISPATCH_DATA_STORAGE_TARGET, side_effect=dispatch_data_storage_tasks)
+    @patch.object(EntityListResource, "PARAMETERS", {"size": 3, "page_size": 10})
+    def test_task_error(self, dispatch_mock):
+        # Prepare test data
+        disappearing_title = "This title should disappear. It is overwritten by new entities"
+        remaining_title = "This title should remain. It is copied from old entities"
+        self.prepare_documents(
+            properties={
+                "2": {"title": disappearing_title},
+                "4": {"title": remaining_title}
+            },
+            task_results={
+                # this task_results should be overwritten and is retried with a failure
+                "1": {"check_doi": {"success": False}},
+                # these task_results should be overwritten and are retried successfully
+                "2": {"check_doi": {"success": False}},
+                "4": {"check_doi": {"success": False}}
+            }
+        )
+        # Execute growth
         with override_settings(TEST_CHECK_DOI_FAILURE_IDENTITIES=["1"]):
             self.dataset.grow(self.entity_type, asynchronous=False)
-        self.assert_initial_grow_failure(dataset_versions=1, collections=1, documents=19, error_documents=1)
+        # Standard assertions
+        # There are four error documents, because three historic documents are not retried
+        # and one new document still fails.
+        self.assert_grow_failure(dataset_versions=2, collections=2, documents=36, error_documents=4)
         self.assert_dataset_output(self.dataset, dataset_versions=1, collections=1, documents=20)
-        dataset_version = DatasetVersion.objects.first()
+        self.assert_dataset_finish()
+        self.assert_document_dispatch(dispatch_mock, ["1", "2", "4"])
+        # Check if expected updates have taken place
+        self.assertEqual(
+            Document.objects.filter(properties__title=disappearing_title).count(), 1,
+            "Expected disappearing title to be preserved in old data, but overwritten in new data."
+        )
+        self.assertEqual(
+            Document.objects.filter(properties__title=remaining_title).count(), 2,
+            "Expected remaining title to be present in old and new data."
+        )
+        dataset_version = self.dataset.get_current_dataset_version()
         self.assertEqual(dataset_version.errors, {
             "tasks": {
                 "check_doi": {
@@ -66,26 +183,197 @@ class TestInitialDatasetReviseStrategy(test_cases.BaseDatasetTestCase):
             },
             "seeding": {}
         })
+
+    @patch(DISPATCH_DATA_STORAGE_TARGET, side_effect=dispatch_data_storage_tasks)
+    def test_growth_retry(self, dispatch_mock):
+        # Set the historic dataset version to ERROR
+        dataset_version = self.dataset.get_current_dataset_version()
+        dataset_version.state = GrowthState.ERROR
+        dataset_version.save()
+        # Prepare test data
+        remaining_title = "This title should remain. New data is never fetched during a retry with historic data"
+        self.prepare_documents(
+            properties={
+                "2": {"title": remaining_title},
+                "4": {"title": remaining_title}
+            },
+            task_results={
+                # these task_results should be overwritten and are retried successfully
+                "1": {"check_doi": {"success": False}},
+                "2": {"check_doi": {"success": False}},
+                "4": {"check_doi": {"success": False}}
+            }
+        )
+        # Execute growth
+        self.dataset.grow(self.entity_type, retry=True, asynchronous=False)
+        # Standard assertions
+        self.assert_grow_success(dataset_versions=1, collections=1, documents=20)
+        self.assert_dataset_output(self.dataset, dataset_versions=1, collections=1, documents=20)
         self.assert_dataset_finish()
+        self.assert_document_dispatch(dispatch_mock, ["1", "2", "4"])
+        # Check if expected updates have taken place
+        self.assertEqual(
+            Document.objects.filter(properties__title=remaining_title).count(), 2,
+            "Expected remaining title to be present in old data."
+        )
+        dataset_version = self.dataset.get_current_dataset_version()
+        self.assertEqual(dataset_version.errors, {
+            "tasks": {
+                "check_doi": {
+                    "fail": 0,
+                    "skipped": 0,
+                    "success": 20
+                }
+            },
+            "seeding": {}
+        })
 
+    @patch(DISPATCH_DATA_STORAGE_TARGET, side_effect=dispatch_data_storage_tasks)
+    def test_growth_retry_completed(self, dispatch_mock):
+        # Completed DatasetVersions should simply trigger a normal growth
+        # Prepare test data
+        remaining_title = "This title should remain. New data is never fetched during a retry with historic data"
+        self.prepare_documents(
+            properties={
+                "2": {"title": remaining_title},
+                "4": {"title": remaining_title}
+            },
+            task_results={
+                # These task_results should be overwritten for new data and get retried
+                "1": {"check_doi": {"success": False}},
+                "2": {"check_doi": {"success": False}},
+                "4": {"check_doi": {"success": False}}
+            }
+        )
+        # Execute growth where retry is not fetching documents, but does create copies and retries tasks
+        self.dataset.grow(self.entity_type, retry=True, asynchronous=False)
+        # Standard assertions
+        self.assert_grow_failure(dataset_versions=2, collections=2, documents=37, error_documents=3)
+        self.assert_dataset_output(self.dataset, dataset_versions=1, collections=1, documents=20)
+        self.assert_dataset_finish()
+        self.assert_document_dispatch(dispatch_mock, ["1", "2", "4"])
+        # Check if expected updates have taken place
+        self.assertEqual(
+            Document.objects.filter(properties__title=remaining_title).count(), 4,
+            "Expected remaining title to be present in old and new data."
+        )
+        dataset_version = self.dataset.get_current_dataset_version()
+        self.assertEqual(dataset_version.errors, {
+            "tasks": {
+                "check_doi": {
+                    "fail": 0,
+                    "skipped": 0,
+                    "success": 20
+                }
+            },
+            "seeding": {}
+        })
 
-class TestContinuationDatasetReviseStrategy(test_cases.BaseDatasetTestCase):
+    @patch(DISPATCH_DATA_STORAGE_TARGET, side_effect=dispatch_data_storage_tasks)
+    @patch.object(EntityListResource, "PARAMETERS", {"size": 3, "page_size": 10})
+    def test_growth_reset(self, dispatch_mock):
+        # When overriding the GrowthStrategy with RESET the new DatasetVersion should ignore historic data
+        # Prepare test data
+        disappearing_title = "This title should disappear. It is overwritten by new entities"
+        self.prepare_documents(properties={
+            "2": {
+                "title": disappearing_title
+            },
+            "4": {
+                "title": disappearing_title
+            }
+        })
+        # Execute growth. Historic data will get ignored
+        self.dataset.grow(self.entity_type, growth_strategy=GrowthStrategy.RESET, asynchronous=False)
+        # Standard assertions
+        self.assert_grow_success(dataset_versions=2, collections=2, documents=23)
+        self.assert_dataset_output(self.dataset, dataset_versions=1, collections=1, documents=3)
+        self.assert_dataset_finish()
+        self.assert_document_dispatch(dispatch_mock, ["0", "1", "2"])
+        # Check if expected updates have taken place
+        self.assertEqual(
+            Document.objects.filter(properties__title=disappearing_title).count(), 2,
+            "Expected disappearing title to be preserved in old data, but missing in new data."
+        )
 
-    dataset_model = Dataset
-    signature = "setting1=const&test"
-    entity_type = "paper"
+    @patch(DISPATCH_DATA_STORAGE_TARGET, side_effect=dispatch_data_storage_tasks)
+    @patch.object(EntityListResource, "PARAMETERS", {"size": 3, "page_size": 10})
+    def test_add_task_definition(self, dispatch_mock):
+        # Set the historic dataset version to PENDING
+        dataset_version = self.dataset.get_current_dataset_version()
+        dataset_version.state = GrowthState.PENDING
+        dataset_version.save()
+        Document.objects.all().update(task_results={}, derivatives={})
+        # Prepare test data
+        remaining_title = "This title should remain. New data is never fetched during a retry with historic data"
+        self.prepare_documents(properties={
+            "2": {
+                "title": remaining_title
+            },
+            "4": {
+                "title": remaining_title
+            }
+        })
+        # Execute growth
+        self.dataset.grow(self.entity_type, retry=True, asynchronous=False)
+        # Standard assertions
+        self.assert_grow_success(dataset_versions=1, collections=1, documents=20)
+        self.assert_dataset_output(self.dataset, dataset_versions=1, collections=1, documents=20)
+        self.assert_dataset_finish()
+        self.assert_document_dispatch(dispatch_mock, [str(ix) for ix in range(0, 20)])  # dispatching all documents
+        # Check if expected updates have taken place
+        self.assertEqual(
+            Document.objects.filter(properties__title=remaining_title).count(), 2,
+            "Expected remaining title to remain during a retry where tasks are added."
+        )
+        dataset_version = self.dataset.get_current_dataset_version()
+        self.assertEqual(dataset_version.errors, {
+            "tasks": {
+                "check_doi": {
+                    "fail": 0,
+                    "skipped": 0,
+                    "success": 20
+                }
+            },
+            "seeding": {}
+        })
 
-    def test_growth_success(self):
-        self.skipTest("not implemented")
-
-    def test_growth_limit(self):
-        self.skipTest("not implemented")
-
-    def test_seeding_error(self):
-        self.skipTest("not implemented")
-
-    def test_growth_retry(self):
-        self.skipTest("not implemented")
-
-    def test_growth_reset(self):
-        self.skipTest("not implemented")
+    @patch(DISPATCH_DATA_STORAGE_TARGET, side_effect=dispatch_data_storage_tasks)
+    @patch.object(EntityListResource, "PARAMETERS", {"size": 3, "page_size": 10})
+    def test_delete_task_definition(self, dispatch_mock):
+        # Set the historic dataset version to PENDING
+        dataset_version = self.dataset.get_current_dataset_version()
+        dataset_version.state = GrowthState.PENDING
+        dataset_version.task_definitions["document"] = {}
+        dataset_version.errors = {
+            "seeding": {},
+            "tasks": {}
+        }
+        dataset_version.save()
+        # Prepare test data
+        remaining_title = "This title should remain. New data is never fetched during a retry with historic data"
+        self.prepare_documents(properties={
+            "2": {
+                "title": remaining_title
+            },
+            "4": {
+                "title": remaining_title
+            }
+        })
+        # Execute growth
+        self.dataset.grow(self.entity_type, retry=True, asynchronous=False)
+        # Standard assertions
+        self.assert_grow_success(dataset_versions=1, collections=1, documents=20)
+        self.assert_dataset_output(self.dataset, dataset_versions=1, collections=1, documents=20)
+        self.assert_dataset_finish()
+        self.assertEqual(dispatch_mock.call_count, 0)
+        # Check if expected updates have taken place
+        self.assertEqual(
+            Document.objects.filter(properties__title=remaining_title).count(), 2,
+            "Expected remaining title to remain during a retry where tasks are added."
+        )
+        dataset_version = self.dataset.get_current_dataset_version()
+        self.assertEqual(dataset_version.errors, {
+            "tasks": {},
+            "seeding": {}
+        })